[Refactor] redis관련 코드 이동 #57
