+++ src/main/java/com/example/starhub/service/UserService.java
@@ -1,5 +1,6 @@
 package com.example.starhub.service;
 
+import com.example.starhub.config.Redis.RedisService;
 import com.example.starhub.dto.request.CreateProfileRequestDto;
 import com.example.starhub.dto.request.CreateUserRequestDto;
 import com.example.starhub.dto.request.UsernameCheckRequestDto;
